Code Refactor! Fixed the Camera_Controller so it follows the code standard and common sense.

+++ Tunnelour/Tunnelour/include/Debug_Data_Display_Controller.h
@@ -170,8 +170,8 @@ class Debug_Data_Display_Controller: public Controller {
   Tileset_Helper::Tileset_Metadata m_debug_tileset_metadata;
   std::string m_debug_metadata_file_path;
   std::vector<Tile_Bitmap*> m_collision_bitmaps;
-  int z_position;
-  int m_bitmap_z_position;
+  float z_position;
+  float m_bitmap_z_position;
   float m_jumping_distance;
   float m_jumping_height;
 };
